Entity, EntityStatus and Player classes replanned

+++ include/EntityStatus.hpp
@@ -9,7 +9,7 @@ public:
     float get_max_health();
     float get_xp();
 
-    int change_health(float value);
-    int change_max_health(float value);
-    int change_xp(float value);
+    float change_health(float value);
+    float change_max_health(float value);
+    float change_xp(float value);
 };
