+++ src/EntityStatus.cpp
@@ -10,15 +10,15 @@ float EntityStatus::get_xp(){
     return this->xp;
 }
 
-int EntityStatus::change_health(float value){
+float EntityStatus::change_health(float value){
     this->health = value;
     return 1;
 }
-int EntityStatus::change_max_health(float value){
+float EntityStatus::change_max_health(float value){
     this->max_health = value;
     return 1;
 }
-int EntityStatus::change_xp(float value){
+float EntityStatus::change_xp(float value){
     this->xp = value;
     return 1;
 }
